VOC report redesign + EFF/VOC parity verified; Jun Excel swap; green tests
- VOC report UI: summary block moved to top, paginated By-Line/Day table
  (5/page), sticky filter bar, hide empty/NA chemical columns toggle,
  green under-consumption heatmap, widened HIGH/LOW rank tables, removed
  Output/VOC(g) columns from By-Line/Day and reconcile pivots
- EFF + VOC formulas audited 1:1 vs June files (no logic change needed):
  EFF KPI = Output*TCT/(DL*WT*3600*Allow); EFF Salary = slot-count based;
  VOC actual=(qty-throw)*factor, allowance=output*recipe (no x1.1)
- Swap May Excel sources for June (Copy of EFF JUNE V6 + structure map)
- salary.html / incentive tweaks; messages i18n keys
- Fix stale tests so full suite is green (407 run, 0 failures, 0 errors):
  add missing VocConsumptionRepository @Mock (DataRetentionServiceTest) and
  LineAssignmentService @MockitoBean (AdminApiControllerTest)

+++ src/main/java/thienloc/manage/dto/SalaryReportDto.java
@@ -30,7 +30,7 @@ public class SalaryReportDto {
         private LocalDate date;
         private double targetQuota;  // planned daily quota (from DailyProductionDto.target)
         private double targetMp;     // standard MP resolved by EFF calc (Excel sheet S col C)
-        private int mp;              // actual MP
+        private double mp;           // actual MP = DLI (Excel sheet S col F = D!I), not DL
         private double wt;
         private double output;       // actual produced quantity
         private String sec;          // e.g. "SEW10"
